c: decompressor: Force integer promotion type for bit_buffer
C does automatic integer promotion when performing calculations with
smaller types (like unsigned integers). But it only guarantees the size
(and type) of an int. But the  bit_buffer is explicitly an uint32_t. It can
therefore be that an int is not 32 bit (depending on the target
architecture).

Also, shifting a signed integer (affecting the sign bit) is undefined
behavior in C. Compilers like GCC can therefore perform optimizations which
break the expected behavior. The shifted type must therefore be forced to
unsigned.

This can be seen when enabling the undefined behavior sanitizer of
GCC/clang:

    decompressor.c:135:43: runtime error: left shift of 131 by 24 places cannot be represented in type 'int'

Fixes: 85ef612bd424 ("decompressor progress")
Signed-off-by: Sven Eckelmann <sven@narfation.org>